requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-utils==0.7.0
uvicorn==0.30.1
bingads==13.0.21.2
google-cloud-bigquery==3.27.0
google-cloud-logging==3.11.3
pydantic==2.7.4
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
setuptools<81
suds-py3==1.4.5.0
requests==2.32.3